requirements.txt
requests
pygments
termcolor
tabulate
pyyaml